remove query methods from User model to repo
