+++ src/app/repositories/user/repository.rs
@@ -1,9 +1,11 @@
+use bson::oid::ObjectId;
 use bson::Document;
+use chrono::Utc;
 use mongodb::bson::{self, doc};
 use mongodb::options::ClientOptions;
 use mongodb::{Client, Collection, Database};
 
-use crate::app::models::user::User;
+use crate::app::models::user::{User, UserProfile};
 use crate::app::services::cache::service::CacheService;
 use crate::app::services::storage::service::StorageService;
 use crate::config::mongodb_config::MongoDBConfig;
@@ -57,5 +59,45 @@ impl UserRepository {
 
     fn get_collection(&self) -> Collection<User> {
         self.storage.get_collection::<User>(&self.collection)
+    }
+
+    pub fn get_update_user_profile_query(user_profile: UserProfile) -> Document {
+        let mut data_to_update = match user_profile {
+            UserProfile::Google(google_profile) => {
+                doc! {
+                    "google.name": google_profile.name,
+                    "google.email": google_profile.email,
+                    "google.email_verified": google_profile.email_verified,
+                    "google.picture": google_profile.picture,
+                }
+            }
+            UserProfile::Facebook(facebook_profile) => {
+                doc! {
+                    "facebook.name": facebook_profile.name,
+                    "facebook.email": facebook_profile.email,
+                }
+            }
+        };
+        data_to_update.insert("updated_at", Utc::now());
+
+        data_to_update
+    }
+    pub fn get_find_user_by_profile_query(user_profile: UserProfile) -> Document {
+        let mut query = doc! {};
+        match user_profile {
+            UserProfile::Google(google_profile) => {
+                query.insert("google.user_id", google_profile.user_id);
+            }
+            UserProfile::Facebook(facebook_profile) => {
+                query.insert("facebook.user_id", facebook_profile.user_id);
+            }
+        }
+
+        query
+    }
+    pub fn get_find_user_by_id_query(user_id: ObjectId) -> Document {
+        doc! {
+            "_id": user_id
+        }
     }
 }
